requirements.txt
PyYAML>=6.0
requests>=2.28
google-auth>=2.0
google-auth-oauthlib>=1.0
google-api-python-client>=2.0
